+++ CHANGELOG.md
@@ -14,8 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Features
 - Sorting filter for name, labels and confirmation states.
 
-### Fixes
-- Rake task for GIS URL.
+### Tasks
+```ruby
+rake gis:populate
+```
 
 ## [1.5.0] - 2022-03-29